pyproject.toml
[project]
name = "pyllmvalidate-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.93.0",
    "gitpython>=3.1.40",
    "python-dotenv>=1.1.1",
]
